Health Check Controller도 수정
