+++ src/main/java/spot/spot/global/healthcheck/command/dto/SampleDto.java
@@ -1,0 +1,11 @@
+package spot.spot.global.healthcheck.command.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor
+public class SampleDto {
+    private int id;
+    private String name;
+}
